refactor(tests): centralized EnvGuard to fix potential undefined behavior during concurrent tests.
