+++ patch_env_guard_autumn_app.patch
@@ -1,0 +1,106 @@
+--- autumn/src/app.rs
++++ autumn/src/app.rs
+@@ -1091,48 +1091,3 @@
+-    use std::sync::{Mutex, OnceLock};
+     use tower::ServiceExt;
++    use crate::test_utils::EnvGuard;
+
+-    pub struct EnvGuard {
+-        _lock: std::sync::MutexGuard<'static, ()>,
+-        previous: Vec<(&'static str, Option<String>)>,
+-    }
+-
+-    impl EnvGuard {
+-        pub fn set_many(entries: &[(&'static str, Option<&str>)]) -> Self {
+-            static ENV_LOCK: OnceLock<Mutex<()>> = OnceLock::new();
+-            let lock = ENV_LOCK
+-                .get_or_init(|| Mutex::new(()))
+-                .lock()
+-                .expect("env mutex poisoned");
+-            let mut previous = Vec::with_capacity(entries.len());
+-            for (key, value) in entries {
+-                previous.push((*key, std::env::var(key).ok()));
+-                match value {
+-                    Some(value) => {
+-                        // SAFETY: test-only helper serializes environment mutation with a process-wide mutex.
+-                        unsafe { std::env::set_var(key, value) };
+-                    }
+-                    None => {
+-                        // SAFETY: test-only helper serializes environment mutation with a process-wide mutex.
+-                        unsafe { std::env::remove_var(key) };
+-                    }
+-                }
+-            }
+-            Self {
+-                _lock: lock,
+-                previous,
+-            }
+-        }
+-    }
+-
+-    impl Drop for EnvGuard {
+-        fn drop(&mut self) {
+-            for (key, previous) in self.previous.iter().rev() {
+-                if let Some(previous) = previous {
+-                    // SAFETY: test-only helper serializes environment mutation with a process-wide mutex.
+-                    unsafe { std::env::set_var(key, previous) };
+-                } else {
+-                    // SAFETY: test-only helper serializes environment mutation with a process-wide mutex.
+-                    unsafe { std::env::remove_var(key) };
+-                }
+-            }
+-        }
+-    }
+--- autumn/src/middleware/dev.rs
++++ autumn/src/middleware/dev.rs
+@@ -202,48 +202,3 @@
+-    use std::sync::{Mutex, OnceLock};
+     use tower::ServiceExt;
++    use crate::test_utils::EnvGuard;
+
+-    struct EnvGuard {
+-        _lock: std::sync::MutexGuard<'static, ()>,
+-        previous: Vec<(&'static str, Option<String>)>,
+-    }
+-
+-    impl EnvGuard {
+-        fn set_many(entries: &[(&'static str, Option<&str>)]) -> Self {
+-            static ENV_LOCK: OnceLock<Mutex<()>> = OnceLock::new();
+-            let lock = ENV_LOCK
+-                .get_or_init(|| Mutex::new(()))
+-                .lock()
+-                .expect("env mutex poisoned");
+-            let mut previous = Vec::with_capacity(entries.len());
+-            for (key, value) in entries {
+-                previous.push((*key, std::env::var(key).ok()));
+-                match value {
+-                    Some(value) => {
+-                        // SAFETY: test-only helper serializes environment mutation with a process-wide mutex.
+-                        unsafe { std::env::set_var(key, value) };
+-                    }
+-                    None => {
+-                        // SAFETY: test-only helper serializes environment mutation with a process-wide mutex.
+-                        unsafe { std::env::remove_var(key) };
+-                    }
+-                }
+-            }
+-            Self {
+-                _lock: lock,
+-                previous,
+-            }
+-        }
+-    }
+-
+-    impl Drop for EnvGuard {
+-        fn drop(&mut self) {
+-            for (key, previous) in self.previous.iter().rev() {
+-                if let Some(previous) = previous {
+-                    // SAFETY: test-only helper serializes environment mutation with a process-wide mutex.
+-                    unsafe { std::env::set_var(key, previous) };
+-                } else {
+-                    // SAFETY: test-only helper serializes environment mutation with a process-wide mutex.
+-                    unsafe { std::env::remove_var(key) };
+-                }
+-            }
+-        }
+-    }
